requirements.txt
pandas==1.5.3
plotly==5.9.0
folium==0.14.0
streamlit==1.25.0
streamlit-folium==0.13.0
inflection==0.5.1
Pillow==9.4.0